Change data type of stages

+++ contracts/sources/libraries/time_lock.move
@@ -16,13 +16,13 @@ module libraries::time_lock {
         deployed_at: u64
     }
 
-    const SRC_WITHDRAWAL: u8 = 0;
-    const SRC_PUBLIC_WITHDRAWAL: u8 = 1;
-    const SRC_CANCELLATION: u8 = 2;
-    const SRC_PUBLIC_CANCELLATION: u8 = 3;
-    const DST_WITHDRAWAL: u8 = 4;
-    const DST_PUBLIC_WITHDRAWAL: u8 = 5;
-    const DST_CANCELLATION: u8 = 6;
+    const SRC_WITHDRAWAL: u64 = 0;
+    const SRC_PUBLIC_WITHDRAWAL: u64 = 1;
+    const SRC_CANCELLATION: u64 = 2;
+    const SRC_PUBLIC_CANCELLATION: u64 = 3;
+    const DST_WITHDRAWAL: u64 = 4;
+    const DST_PUBLIC_WITHDRAWAL: u64 = 5;
+    const DST_CANCELLATION: u64 = 6;
 
     // To create timelock
     public fun create_timelock (src_withdrawal: u64,src_public_withdrawal: u64,src_cancellation: u64,src_public_cancellation: u64,
@@ -70,10 +70,10 @@ module libraries::time_lock {
     public fun get_dst_cancellation(timelock: &Timelocks): u64 {
         timelock.dst_cancellation
     }
-    
+
 
     // Get time when a stage becomes active
-    public fun get(timelock: &Timelocks, stage: u8): u64{
+    public fun get(timelock: &Timelocks, stage: u64): u64{
         let offset = if (stage == SRC_WITHDRAWAL){
             timelock.src_withdrawal
         } else if (stage == SRC_PUBLIC_WITHDRAWAL){
